refactor: usa BigDecimal nos valores financeiros

+++ src/main/java/br/com/itau/geradornotafiscal/service/strategy/AliquotaTributariaStrategy.java
@@ -1,10 +1,11 @@
 package br.com.itau.geradornotafiscal.service.strategy;
 
 import br.com.itau.geradornotafiscal.model.Destinatario;
+import java.math.BigDecimal;
 
 public interface AliquotaTributariaStrategy {
 
     boolean suporta(Destinatario destinatario);
 
-    double calcularAliquota(double valorTotalItens);
+    BigDecimal calcularAliquota(BigDecimal valorTotalItens);
 }
