+++ src/main/java/br/com/itau/geradornotafiscal/service/strategy/FreteStrategy.java
@@ -1,10 +1,11 @@
 package br.com.itau.geradornotafiscal.service.strategy;
 
 import br.com.itau.geradornotafiscal.model.Regiao;
+import java.math.BigDecimal;
 
 public interface FreteStrategy {
 
     Regiao regiao();
 
-    double calcular(double valorFrete);
+    BigDecimal calcular(BigDecimal valorFrete);
 }
